refactor: To improve user experience, it's better to have a single Scriptable Renderer Feature that enquees two passes (OutputOutlineTexturesPass, and OutlineRenderPass)
Since having two Scriptable Renderer Features (SRF) would requires that the users add in a specific order the SRFs (OutputOutlineTexturesRendererFeature, and then, the OutlineRendererFeature), this could be prone error.

With this refactor, there is only one SRF that sets up (enqueues) the Render Passes in a specific order without user intervention.